Move functions out of util only used in ruby_linter
